refactor(lean): consolidate shared contracts
