+++ src/tools/lean/leanTypes.ts
@@ -1,11 +1,81 @@
 /**
- * VS Code-free types and pure functions shared between
- * the Lean tool implementations and the VS Code integration layer.
+ * VS Code-free contracts and helpers shared by the Lean tool implementations
+ * and host integration layers.
  */
 
+// Local imports - utils
 import type { GenericDiagnostic } from '@utils/diagnostics/diagnosticFormatting';
 
-import type { LspHover } from './lspTypes';
+// Constants and commands
+
+/** Marketplace ID of the Lean 4 VS Code extension. */
+export const LEAN4_EXTENSION_ID = 'leanprover.lean4';
+
+/** Per-file commands surfaced by `lean_file`. */
+export const LEAN_FILE_COMMANDS = ['restart', 'refresh_dependencies'] as const;
+export type LeanFileCommand = (typeof LEAN_FILE_COMMANDS)[number];
+
+/** Project-scope commands surfaced by `lean_project`. */
+export const LEAN_PROJECT_COMMANDS = [
+  // Server commands
+  'restart_server',
+  'stop_server',
+  // Project commands
+  'build',
+  'clean',
+  'fetch_cache',
+  'fetch_file_cache',
+  // Setup commands
+  'install_elan',
+  'install_deps',
+  'update_elan',
+  'select_toolchain',
+] as const;
+export type LeanProjectCommand = (typeof LEAN_PROJECT_COMMANDS)[number];
+
+/** Human-readable label for each server mode (used in the dashboard surface). */
+export const LEAN_SERVER_MODE_LABELS = {
+  'vscode-extension': LEAN4_EXTENSION_ID,
+  'direct-lsp': 'direct LSP',
+} as const;
+
+// LSP contracts
+
+type LspMarkupContent = {
+  kind: 'plaintext' | 'markdown' | string;
+  value: string;
+};
+
+type LspMarkedString = string | { language: string; value: string };
+
+export interface LspHover {
+  contents: LspMarkedString | LspMarkupContent | LspMarkedString[];
+  range?: LspRange;
+}
+
+export interface LspDiagnostic {
+  range: LspRange;
+  message: string;
+  severity?: number;
+  source?: string;
+}
+
+export interface LspPublishDiagnosticsParams {
+  uri: string;
+  diagnostics: LspDiagnostic[];
+}
+
+interface LspRange {
+  start: LspPosition;
+  end: LspPosition;
+}
+
+interface LspPosition {
+  line: number;
+  character: number;
+}
+
+// Lean tool contracts
 
 /** Response from $/lean/plainGoal LSP request */
 export interface PlainGoal {
@@ -31,6 +101,8 @@ export interface LspResult<T> {
 export interface LeanDiagnostic extends GenericDiagnostic {
   source?: string;
 }
+
+// Helpers
 
 /** Extract text value from hover contents (handles all LSP content formats) */
 export function extractHoverText(
